requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
djangorestframework==3.12.2
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.8.0
pylint==2.6.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
